requirements.txt
streamlit
langchain_community
pypdf
langchain-ollama
langchain-chroma
pandas